atualizando seeds

+++ db/seeds.rb
@@ -1,20 +1,17 @@
-# This file should ensure the existence of records required to run the application in every environment (production,
-# development, test). The code here should be idempotent so that it can be executed at any point in every environment.
-# The data can then be loaded with the bin/rails db:seed command (or created alongside the database with db:setup).
-#
-# Example:
-#
-#   ["Action", "Comedy", "Drama", "Horror"].each do |genre_name|
-#     MovieGenre.find_or_create_by!(name: genre_name)
-#   end
-puts "cleaning up database"
-Product.destroy_all
-puts "database cleaned"
-products = [
-  {name: "Guylian", description: "Belgian Chocolate", price: 8.5, stock: 10},
-  {name: "Lindt", description: "Swiss Chocolate", price: 7.5, stock: 8},
-  {name: "Baci", description: "Italian Chocolate", price: 3.5, stock: 12},
-  {name: "Tony's", description: "American Chocolate", price: 5.5, stock: 9},
-  {name: "Novi", description: "Italian Chocolate", price: 4.5, stock: 7}
-  ]
-puts "products created"
+if Rails.env.development?
+  puts "Cleaning up database"
+  User.destroy_all
+  Product.destroy_all
+  puts "Database cleaned"
+
+  admin = User.create!(email: "admin@chocoshop.com", password: ENV["ADMIN_PASSWORD"])
+  Product.create!(name: "Guylian", description: "Belgian Chocolate",price: 8.5, stock: 10, user: admin)
+  Product.create!(name: "Lindt", description: "Swiss Chocolate", price: 7.5, stock: 8, user: admin)
+  Product.create!(name: "Baci", description: "Italian Chocolate", price: 3.5, stock: 12, user: admin)
+  Product.create!(name: "Tony's", description: "American Chocolate", price: 5.5, stock: 9, user: admin)
+  puts "Products created"
+end
+
+if Rails.env.production?
+  # Criar um seed de production caso necessário
+end
